STRUCT: Move around files to flatten structure
